fix: occasional 500 error when being scraped for metrics
We are seeing errors on the /metrics end-point suggesting that the same metric is attempted to being exported more than once during the same collect.

I suspect this is because the ttlcache.Range() function doesn't grab a mutex for the entire loop, and that the cleanup or other modication is causing the same values to be emitted more than once during the same iteration.

This change gets a copy of the map then iterates over that, and hopefully stops that error condition from manifesting itself.