feat: add tail_pihole option for reverse_dns enricher
After running with reverse_dns for a while we noted the results weren't that interesting. The result for many IPs is a generic load-balancer or other IP at various cloud providers.

In our case, we are running pihole on the same server as this process, so this adds an option to tail the pihole log, and use that to see the original query before CNAME traversals, then cache the final IP resolved for a given initial query.

This may give more interesting results.